Modified MultiLanguage plugin's tests. Moved multi language related entities tests in multilanguageplugintests.cs and deleted multilanguagerelatedplugintests.cs Modified the tests to work with the new logic of using a prefix tag in the opc_name value.
