+++ src/Compliance.Plugins.Tests/MultiLanguagePluginTests.cs
@@ -9,6 +9,8 @@ namespace Compliance.Plugins.Tests
 {
     public class MultiLanguagePluginTests
     {
+        private static readonly string prefix = "|^|";
+
         public class when_creating_multilanguageitem
         {
             public opc_theme GetMockedMultiLanguageItem()
@@ -50,8 +52,8 @@ namespace Compliance.Plugins.Tests
                 var context = new XrmFakedContext();
                 var pluginContext = context.GetDefaultPluginContext();
 
-                var expectedName = "Technology|Technologie";
                 var theme = GetMockedMultiLanguageItem();
+                var expectedName = $"{prefix}{theme.opc_nameenglish}|{theme.opc_namefrench}";
 
                 var inputs = new ParameterCollection { { "Target", theme } };
 
@@ -152,8 +154,8 @@ namespace Compliance.Plugins.Tests
                 var context = new XrmFakedContext();
                 var pluginContext = context.GetDefaultPluginContext();
 
-                var expectedName = "TechnoEng|TechnoFra";
                 var multiLanguageEntity = GetMockedMultiLanguageEntity();
+                var expectedName = $"{prefix}{multiLanguageEntity.opc_nameenglish}|{multiLanguageEntity.opc_namefrench}";
 
                 var inputs = new ParameterCollection { { "Target", multiLanguageEntity } };
 
@@ -165,11 +167,6 @@ namespace Compliance.Plugins.Tests
 
                 // Assert
                 multiLanguageEntity.opc_name.Should().Be(expectedName);
-
-                // TODO: Or is it better this way?
-                multiLanguageEntity.opc_name.Should().StartWith(multiLanguageEntity.opc_nameenglish);
-                multiLanguageEntity.opc_name.Should().EndWith(multiLanguageEntity.opc_namefrench);
-                multiLanguageEntity.opc_name.Should().Contain("|");
             }
 
             [Fact]
@@ -221,7 +218,7 @@ namespace Compliance.Plugins.Tests
                 return new opc_theme
                 {
                     Id = Guid.NewGuid(),
-                    opc_name = "Technology|Technologie",
+                    opc_name = "|^|Technology|Technologie",
                     opc_nameenglish = "Technology",
                     opc_namefrench = "Technologie"
                 };
@@ -234,8 +231,8 @@ namespace Compliance.Plugins.Tests
                 var context = new XrmFakedContext();
                 var pluginContext = context.GetDefaultPluginContext();
 
-                var expectedName = "Technologie";
                 var multiLanguageEntity = GetMockedMultiLanguageEntity();
+                var expectedName = multiLanguageEntity.opc_namefrench;
 
                 var outputs = new ParameterCollection { { "BusinessEntity", multiLanguageEntity } };
 
@@ -257,8 +254,8 @@ namespace Compliance.Plugins.Tests
                 var context = new XrmFakedContext();
                 var pluginContext = context.GetDefaultPluginContext();
 
-                var expectedName = "Technology";
                 var multiLanguageEntity = GetMockedMultiLanguageEntity();
+                var expectedName = multiLanguageEntity.opc_nameenglish;
 
                 var outputs = new ParameterCollection { { "BusinessEntity", multiLanguageEntity } };
 
@@ -280,8 +277,8 @@ namespace Compliance.Plugins.Tests
             {
                 EntityCollection entityCollectionThemes = new EntityCollection();
 
-                entityCollectionThemes.Entities.Add(new opc_theme { Id = Guid.NewGuid(), opc_nameenglish = "Technology", opc_namefrench = "Technologie", opc_name = "Technology|Technologie" });
-                entityCollectionThemes.Entities.Add(new opc_theme { Id = Guid.NewGuid(), opc_nameenglish = "Public Services", opc_namefrench = "Services Publiques", opc_name = "Public Services|Services Publiques" });
+                entityCollectionThemes.Entities.Add(new opc_theme { Id = Guid.NewGuid(), opc_nameenglish = "Technology", opc_namefrench = "Technologie", opc_name = "|^|Technology|Technologie" });
+                entityCollectionThemes.Entities.Add(new opc_theme { Id = Guid.NewGuid(), opc_nameenglish = "Public Services", opc_namefrench = "Services Publiques", opc_name = "|^|Public Services|Services Publiques" });
 
                 return entityCollectionThemes;
             }
@@ -331,6 +328,139 @@ namespace Compliance.Plugins.Tests
                 foreach (opc_theme multiLanguageEntity in multiLanguageEntityCollection.Entities)
                 {
                     multiLanguageEntity.opc_name.Should().Be(multiLanguageEntity.opc_nameenglish);
+                }
+            }
+        }
+
+        public class when_retrieving_multilanguage_relateditem
+        {
+            public opc_topic GetMockedMultiLanguageEntity()
+            {
+                return new opc_topic
+                {
+                    Id = Guid.NewGuid(),
+                    opc_name = "|^|Artificial Intelligence|Intelligence Artificielle",
+                    opc_nameenglish = "Artificial Intelligence",
+                    opc_namefrench = "Intelligence Artificielle",
+                    opc_themeid = new EntityReference() { Name = "|^|Technology|Technologie" }
+                };
+            }
+
+            [Fact]
+            public void opc_name_should_be_french_when_ui_is_french()
+            {
+                // Arrange
+                var context = new XrmFakedContext();
+                var pluginContext = context.GetDefaultPluginContext();
+                var multiLanguageEntity = GetMockedMultiLanguageEntity();
+                var outputs = new ParameterCollection { { "BusinessEntity", multiLanguageEntity } };
+                var expectedName = "Technologie";
+
+                pluginContext.OutputParameters = outputs;
+                pluginContext.MessageName = "Retrieve";
+                pluginContext.SharedVariables.Add("UserLocaleId", 1036);
+
+                // Act
+                context.ExecutePluginWith<MultiLanguagePlugin>(pluginContext);
+
+                // Assert
+                multiLanguageEntity.opc_themeid.Name.Should().Be(expectedName);
+            }
+
+            [Fact]
+            public void opc_name_should_be_english_when_ui_is_english()
+            {
+                // Arrange
+                var context = new XrmFakedContext();
+                var pluginContext = context.GetDefaultPluginContext();
+                var multiLanguageEntity = GetMockedMultiLanguageEntity();
+                var outputs = new ParameterCollection { { "BusinessEntity", multiLanguageEntity } };
+                var expectedName = "Technology";
+
+                pluginContext.OutputParameters = outputs;
+                pluginContext.MessageName = "Retrieve";
+                pluginContext.SharedVariables.Add("UserLocaleId", 1033);
+
+                // Act
+                context.ExecutePluginWith<MultiLanguagePlugin>(pluginContext);
+
+                // Assert
+                multiLanguageEntity.opc_themeid.Name.Should().Be(expectedName);
+            }
+        }
+
+        public class when_retrieving_multiple_multilanguage_relateditem
+        {
+            public EntityCollection GetMultiLanguageEntityCollection()
+            {
+                EntityCollection entityCollectionTopics = new EntityCollection();
+
+                entityCollectionTopics.Entities.Add(new opc_topic
+                {
+                    Id = Guid.NewGuid(),
+                    opc_name = "|^|Artificial Intelligence|Intelligence Artificielle",
+                    opc_nameenglish = "Artificial Intelligence",
+                    opc_namefrench = "Intelligence Artificielle",
+                    opc_themeid = new EntityReference() { Name = "|^|Technology|Technologie" },
+                    opc_theme_topics_themeid = new opc_theme { opc_nameenglish = "Technology", opc_namefrench = "Technologie" }
+                });
+
+                entityCollectionTopics.Entities.Add(new opc_topic
+                {
+                    Id = Guid.NewGuid(),
+                    opc_name = "|^|The Service|Le Service",
+                    opc_nameenglish = "The Service",
+                    opc_namefrench = "Le Service",
+                    opc_themeid = new EntityReference() { Name = "|^|Public Services|Services Publiques" },
+                    opc_theme_topics_themeid = new opc_theme { opc_nameenglish = "Public Services", opc_namefrench = "Services Publiques" }
+                });
+
+                return entityCollectionTopics;
+            }
+
+            [Fact]
+            public void opc_names_should_be_french_when_ui_is_french()
+            {
+                // Arrange
+                var context = new XrmFakedContext();
+                var pluginContext = context.GetDefaultPluginContext();
+                var multiLanguageEntityCollection = GetMultiLanguageEntityCollection();
+                var outputs = new ParameterCollection { { "BusinessEntityCollection", multiLanguageEntityCollection } };
+
+                pluginContext.OutputParameters = outputs;
+                pluginContext.MessageName = "RetrieveMultiple";
+                pluginContext.SharedVariables.Add("UserLocaleId", 1036);
+
+                // Act
+                context.ExecutePluginWith<MultiLanguagePlugin>(pluginContext);
+
+                // Assert
+                foreach (opc_topic multiLanguageEntity in multiLanguageEntityCollection.Entities)
+                {
+                    multiLanguageEntity.opc_themeid.Name.Should().Be(multiLanguageEntity.opc_theme_topics_themeid.opc_namefrench);
+                }
+            }
+
+            [Fact]
+            public void opc_names_should_be_english_when_ui_is_english()
+            {
+                // Arrange
+                var context = new XrmFakedContext();
+                var pluginContext = context.GetDefaultPluginContext();
+                var multiLanguageEntityCollection = GetMultiLanguageEntityCollection();
+                var outputs = new ParameterCollection { { "BusinessEntityCollection", multiLanguageEntityCollection } };
+
+                pluginContext.OutputParameters = outputs;
+                pluginContext.MessageName = "RetrieveMultiple";
+                pluginContext.SharedVariables.Add("UserLocaleId", 1033);
+
+                // Act
+                context.ExecutePluginWith<MultiLanguagePlugin>(pluginContext);
+
+                // Assert
+                foreach (opc_topic multiLanguageEntity in multiLanguageEntityCollection.Entities)
+                {
+                    multiLanguageEntity.opc_themeid.Name.Should().Be(multiLanguageEntity.opc_theme_topics_themeid.opc_nameenglish);
                 }
             }
         }
